# temp_openwa/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/integration/plugins/{pluginId}/instances": {
      "post": {
        "operationId": "IntegrationInstanceController_create",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Instance created. The plaintext ingress secret and verifyToken are revealed once in this response — store them immediately (both masked on every later read).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceView"
                }
              }
            }
          }
        },
        "tags": [
          "integration"
        ]
      },
      "get": {
        "operationId": "IntegrationInstanceController_list",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instances for the plugin (secrets masked).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstanceView"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "integration"
        ]
      }
    },
    "/api/integration/plugins/{pluginId}/instances/{instanceId}": {
      "get": {
        "operationId": "IntegrationInstanceController_getOne",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The instance (secret masked).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceView"
                }
              }
            }
          }
        },
        "tags": [
          "integration"
        ]
      },
      "patch": {
        "operationId": "IntegrationInstanceController_patch",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instance updated (secret masked).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceView"
                }
              }
            }
          }
        },
        "tags": [
          "integration"
        ]
      },
      "delete": {
        "operationId": "IntegrationInstanceController_remove",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Instance deleted and its session scope torn down."
          }
        },
        "tags": [
          "integration"
        ]
      }
    },
    "/api/integration/plugins/{pluginId}/instances/{instanceId}/regenerate-secret": {
      "post": {
        "operationId": "IntegrationInstanceController_regenerate",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Secret regenerated. The new plaintext secret is revealed once in this response; the verifyToken is also shown (unchanged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceView"
                }
              }
            }
          }
        },
        "tags": [
          "integration"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.12.4",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "OpenWA instance",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
...
      "CreateInstanceDto": {
        "type": "object",
        "properties": {
          "instanceId": {
            "type": "string",
            "description": "Operator-chosen instance id (unique within the plugin). Namespaces the ingress URL and the instance secret.",
            "example": "chatwoot-prod-1"
          },
          "sessionScope": {
            "type": "string",
            "description": "Session id the instance is scoped to. Omit for all sessions.",
            "example": "8f3c2b1a-9d4e-4c7a-8b2f-1e6d5a4c3b2a"
          },
          "verifyToken": {
            "type": "string",
            "description": "Token echoed back for the provider webhook verification handshake. Auto-generated when omitted.",
            "example": "a1b2c3d4e5f6"
          },
          "secret": {
            "type": "string",
            "description": "Ingress HMAC secret shared with the provider. Omit to auto-generate a random 64-hex secret. Masked (****) on every read.",
            "writeOnly": true,
            "example": "super-secret-provider-webhook-key"
          },
          "config": {
            "type": "object",
            "description": "Per-instance config slice passed to the adapter (shape defined by the plugin).",
            "example": {
              "apiKey": "chatwoot-key",
              "inboxId": 42
            }
          }
        },
        "required": [
          "instanceId"
        ]
      },
      "IngressUrl": {
        "type": "object",
        "properties": {
          "route": {
            "type": "string",
            "description": "Plugin-declared route segment the provider posts to."
          },
          "url": {
            "type": "string",
            "description": "Full ingress URL (BASE_URL + plugin/instance/route), or a relative path when BASE_URL is unset."
          }
        },
        "required": [
          "route",
          "url"
        ]
      },
      "InstanceView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Instance row id."
          },
          "pluginId": {
            "type": "string",
            "description": "Plugin id this instance belongs to."
          },
          "instanceId": {
            "type": "string",
            "description": "Operator-chosen instance id (unique within the plugin)."
          },
          "sessionScope": {
            "type": "object",
            "description": "Session id the instance is scoped to, or null for all sessions.",
            "nullable": true
          },
          "secret": {
            "type": "string",
            "description": "Ingress HMAC secret. Masked ('***') on every read; plaintext returned only once on create/regenerate-secret."
          },
          "verifyToken": {
            "type": "object",
            "description": "Provider verify-token. Masked ('***') on reads when set; plaintext on create/regenerate-secret.",
            "nullable": true
          },
          "config": {
            "type": "object",
            "description": "Per-instance config slice passed to the adapter, or null. Fields flagged `secret` in the plugin's config schema are masked ('***') on EVERY response — including create/regenerate-secret; only the ingress secret and verifyToken are ever revealed once.",
            "nullable": true
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether ingress is accepted and dispatch is active."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp."
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp."
          },
          "ingressUrls": {
            "description": "Ingress URLs the provider posts webhook deliveries to.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IngressUrl"
            }
          }
        },
        "required": [
          "id",
          "pluginId",
          "instanceId",
          "secret",
          "enabled",
          "createdAt",
          "updatedAt",
          "ingressUrls"
        ]
      },
      "UpdateInstanceDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Whether the instance is enabled (ingress accepted, dispatch active).",
            "example": true
          },
          "sessionScope": {
            "type": "string",
            "description": "Session id the instance is scoped to. Omit for all sessions."
          },
          "config": {
            "type": "object",
            "description": "Per-instance config slice passed to the adapter."
          }
        }
      }
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}